+++ tests/unit/EmojiTest.php
@@ -1,13 +1,11 @@
 <?php
 
-use Foodsharing\Lib\Db\Db;
 use Foodsharing\Modules\Core\Database;
 
 class EmojiTest extends \Codeception\Test\Unit
 {
     protected UnitTester $tester;
     private Database $db;
-    private Db $db2;
 
     private $user1;
     private $user2;
@@ -18,7 +16,6 @@ class EmojiTest extends \Codeception\Test\Unit
     protected function _before()
     {
         $this->db = $this->tester->get(Database::class);
-        $this->db2 = $this->tester->get(Db::class);
         $this->user1 = $this->tester->createFoodsharer();
         $this->user2 = $this->tester->createFoodsharer();
         $this->conversation = $this->tester->createConversation([
@@ -37,12 +34,6 @@ class EmojiTest extends \Codeception\Test\Unit
     public function testEmojiHandlingWithPDO()
     {
         $body = $this->db->fetchValueByCriteria('fs_msg', 'body', ['id' => $this->messageId]);
-        $this->assertEquals($this->messageBody, $body);
-    }
-
-    public function testEmojiHandlingWithMysqli()
-    {
-        $body = $this->db2->qOne('select body from fs_msg where id = ' . $this->messageId);
         $this->assertEquals($this->messageBody, $body);
     }